+++ src/LuaObject.hpp
@@ -36,7 +36,7 @@ class LuaObject : public RefCounted {
 
 public:
 	virtual const sol::reference& get_lua_object() const;
-	LuaState *get_lua_state() const;
+	Ref<LuaState> get_lua_state() const;
 
 	uint64_t get_pointer_value() const;
 
@@ -70,18 +70,14 @@ public:
 			ERR_FAIL_MSG("FIXME: invalid reference to Lua object");
 		}
 		known_objects.insert(lua_object.pointer(), this);
-		if (LuaState *state = get_lua_state()) {
-			lua_state = Ref(state);
-		}
+		lua_state = get_lua_state();
 	}
 	LuaObjectSubclass(const TReference& lua_object) : lua_object(lua_object) {
 		if (!lua_object.valid()) {
 			ERR_FAIL_MSG("FIXME: invalid reference to Lua object");
 		}
 		known_objects.insert(lua_object.pointer(), this);
-		if (LuaState *state = get_lua_state()) {
-			lua_state = Ref(state);
-		}
+		lua_state = get_lua_state();
 	}
 
 	virtual ~LuaObjectSubclass() {